aws_tutorials/requirements.txt
aws-cdk.core==1.204.0
aws-cdk.aws_iam==1.204.0
aws-cdk.aws_s3==1.204.0
awscli==1.32.17
black==23.12.1
boto3==1.34.17
flake8==7.0.0
jupyter==1.0.0
matplotlib==3.4.3
nbqa==1.7.1
numpy==1.26.3
pandas==2.1.4
seaborn==0.11.2
statsmodels